E-production model draft
look on view "e-production model1"
there are two stock/flow models used as measurement/output variables:
- GHG emission rate
- energy stock (the total amount of energy in the system before it's being used by households, transport and others)

GHG emissions are created from Households and Transportation as well as E-production
All three sectors also demand energy to work. In general this can be be energy in different forms. The aim of this work is to eliminate energy in fossil form and convert all energy demand into electricity.

Conventional e-production is done by converting fossil resources into electricity based on the capacity of the plants. Mining of resources as well as construction of plants (conventional and unconventional/renewable) requires also energy and other resources, as well as all the different processes have some sort of emission.